Initial plugin version, with data dump from old bot
Created this basic plugin on request, using the "database" provided by
the requester from an older bot that has become increasingly unreliable.
We'll be able to always have at least one "waifu" command available,
even if the "real" (original) one is down.

Thanks to Peorth for the data file, and to the maintainer of imouto-baka
on Rizon IRC for building the original bot.