Mostly revert 563a22f to work with revised F/GO servant marking method
IRC people decided that 'tmwaifu' wasn't as good of a command, so the
waifu list has been updated again with '(F/GO)' in the name of each
character that should be included in the 'fgowaifu' choices.

Best of both worlds: We can have a unified TYPE-MOON universe in the
main 'waifu' command, *and* only output F/GO servants for 'fgowaifu'.