Review/fix "J" series
I guess Joshiraku is a *little* messy now with the multiple nicknames,
but for everyone other than Kukuru the puns are pretty fun and obvious.

Without having seen the show, I couldn't put together a good nickname
for Kukuru. Her pun origin is a lot less clear-cut than the others.