Review/fix remaining "K" series, from "Katanagari" onward
Despite what the last commit says, I didn't actually change anything in
the "Katanagari" section until now. Found a partial edit that Sublime
Text remembered (but that I had not actually saved to the file).

Biggest change here is in "K-ON!", which has been punctuated and moved
to the top of all "K" series because hyphen comes before any letters in
the ASCII table order. And of course, every single character in the
franchise was touched (name order & nicknames), plus many added.