Check/fix "Z" series
It's literally just Zero no Tsukaima (well, four seasons of it).

Shows with lots of Western name order are annoying because of my foolish
decision to sort characters by surname, which is very convenient in most
cases (F9 on the range of lines in Sublime Text just works) but requires
manual work when dealing with non-Eastern names. In this case, since ALL
of the characters are in Western order, I just sorted by given name.

No one else will ever notice, anyway.