Add Yuri Is My Job! characters
I don't mind this being the only thing I add today. It was complicated
to collect all the names because it's a manga with no anime adaptation,
so it automatically gets MUCH less attention on MAL and the character
listing is incomplete.

Collated data also from Fandom (ew), aniSearch, and Wikipedia.