Add Do It Yourself!! characters
Currently-airing show with no source material (original), so adding it
now is probably creating more work for myself in the future tbh. It'll
have to be revisted and updated when MAL's data is more fleshed out
after this anime season ends.

Yes, I looked for an outside wiki/source and couldn't find one with more
details than MAL already had.