Revamp BanG Dream! franchise from AniDB data
I'll probably go back on this later, but for now we're using the idea
that this is primarily an *anime* waifu list literally, and putting the
characters from Bandori in according to when they first appeared OR
which series first gave them a main role. (Admittedly I was not very
consistent about that. Did most of this on my phone, on the highway,
because what else can you do on a road trip?)