Add Juuni Kokuki characters
I fudged this one a bit. Some of the characters had names that sounded
like not-a-name, or were obviously elderly, etc. and not suitable for
inclusion in a waifu list.

It's kind of surprising how flaky some of the data for older shows is on
AniDB. I figured the older something is, the more likely it is to have
complete data.