Write waifu suggestions directly to a file
I decided that the complexity of caching suggestions to `bot.db` isn't
worth it. Savings on SSD wear are likely to be negligible as this
feature doesn't see heavy use, and it's actually *more* likely to lose
data that's only cached in `bot.memory` if the bot crashes or otherwise
fails to shut down gracefully. Plus, plugin values in the database will
be truncated to a fairly short length on certain DB backends (see issue
tracking this upstream, sopel-irc/sopel#2291).

Tested live adding a few waifus, viewed the file (making sure newlines
were handled correctly), and then used `.clearwaifus` to delete it.