Rework Fatal Fury data
No doubt Fatal Fury 2 was in the list as a game. Apparently *I* added
this as part of a Dead or Alive cleanup...? (So says `git blame`.)

Anyway, the same character appears in anime based on the game franchise,
so that's what I will include here--along with some other related titles
found in AniDB data.