Tweak [Locodol] entries
This didn't include enough changes to qualify as a "Revamp", IMO. Added
a couple of missing minor characters and modified the order, mostly.

Mikoze Kana is not mentioned ANYWHERE on AniDB but she *is* mentioned by
other sites talking about the franchise and has a VA (Kusuda Aina, no
less) so I don't want to remove her from the list.

Still wish JSON supported comments. Still thinking of a switch to JSON5.