"Fix" Fuuto Tantei characters (with a giant asterisk)
This one is a big mess. AniDB and MAL have *nearly* disjoint sets of
characters, so trying to standardize on one or the other was a flop.
(AniDB has not even enough characters to trigger the "Show All" button
in its Cast section.)

I followed my nose to the Kamen Rider wiki on Fandom (yes, ew), where
yet another overlapping-but-different list of characters was presented.

By now, I've somewhat given up on getting this one to be Correct™ unless
an actual KR fan comes along to put the list together properly, because
as a layman it's really hard to know who's "real" or not without going
and watching the anime myself with a notepad.