Update Natsu-iro Kiseki
Sometimes I dislike how AniDB presents data. This is one of those times.

On MAL, the parent/sibling characters are listed by name, with relation
type in the character's bio.

On AniDB, *most* (but NOT ALL 😤) parent/sibling characters are listed
as "[MainCharGivenName] no [Relation]".

If I had to guess, AniDB's listing probably reflects the show's screen
credits. It's exceedingly boring, though.

Therefore, I merged the two sources. Parents and siblings are listed by
name *and* relation to one of the main characters. We will see if this
presentation gets annoying in practice…