Use `include_admins` in `@plugin.rate_user` if available
Sopel 8.1 hasn't released yet with the new feature. Even if it had, that
wouldn't be a big enough reason to drop Sopel 8.0 immediately.

This trick creates a custom decorator, using the new kwarg if available
to include admins in rate limiting, and omitting it if not.

I think the `inspect` module is less faff here than constructing a
couple of proper Version objects (one from `sopel.__version__`) for
comparison. For better or worse, Sopel isn't *amazingly* well set up to
do a runtime version check; like in web JavaScript, it's better to just
check directly for wanted features.