Update to Zenless Zone Zero characters as of 2025-10-14
Added newly announced and leaked characters to the "Upcoming characters"
category. Placed some (previously) upcoming characters into the main
("Playable"?) category (while fixing up their names). Moved "Colonel
Isolde" to her place in the "Important NPCs" category since she won't be
playable. Fixed some erroneous information about `Belle`; she *is* in
the agent list, but her number is just very high (2011-2036 [I'm unsure
which exact number she is so don't ask!]).

*Special consideration: Magus Magnusson*

Magus Magnusson is the downloaded mind of the real (but now deceased)
Magus into the talking, fire-breathing robotic tail attached to Orpheus
(who also happens to be a clone of the real Magus' body) complete with
her own VA. She is also the captain of the Obol Squad: Soldier 11,
Trigger, Orphie and Seed.

There is a lot of community fanart and comics of the MC having NTR
situations with Magus and Orphie which is why I think she deserves to be
added to this list.