Add xxxHOLiC franchise
Well, to the best of my ability to disentangle it from the other related
properties. It's entwined with Tsubasa Chronicle's setting, which links
to several other shows and just gets overwhelming.

So instead, I went through xxxHOLiC by itself, and will do the same for
the others when they come up.